Add Entry struct and methods for stream entries
